Move CreateTaskUseCase to create folder
