+++ apps/android/app/src/main/kotlin/com/taskodoro/android/app/tasks/create/CreateTaskViewModel.kt
@@ -20,7 +20,7 @@ import androidx.annotation.StringRes
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.taskodoro.android.R
-import com.taskodoro.tasks.CreateTaskUseCase
+import com.taskodoro.tasks.create.CreateTaskUseCase
 import com.taskodoro.tasks.model.Task
 import com.taskodoro.tasks.validator.TaskValidatorError
 import com.taskodoro.tasks.validator.ValidatorError
