+++ taskodoro/src/commonTest/kotlin/com/taskodoro/tasks/create/CreateTaskUseCaseTest.kt
@@ -14,9 +14,10 @@
  *    limitations under the License.
  */
 
-package com.taskodoro.tasks
+package com.taskodoro.tasks.create
 
 import com.taskodoro.helpers.anyTask
+import com.taskodoro.tasks.TaskRepository
 import com.taskodoro.tasks.model.Task
 import com.taskodoro.tasks.validator.TaskValidatorError
 import com.taskodoro.tasks.validator.Validator
